feat(theme): add the ability to add (even partial) background and foreground to color object
While declaring the primary, accent and warn via the color object is easy and straight-forward, the configuration of foreground and background proves rather tedious.
This commit adds the ability to declare foreground and background in the color object when defining a theme, even partially, meaning that missing values are taken from the default light or dark theme respectively.

This commit fixes some possible NPEs in _theming.scss

Fixes #11390, #6244